Minimum Deletions to Make Array Divisible
Problem Link : https://leetcode.com/problems/minimum-deletions-to-make-array-divisible/description/

Approach : Firstly i computed the gcd of all the elements in numsDivide array and then checked for the smallest divisor of this obtained element is present or not in the nums array. Then accordingly returned the result as i if it is matched otherwise returned 0.