Shortest Unsorted Continuous Subarray Medium
Problem Link : https://leetcode.com/problems/shortest-unsorted-continuous-subarray/description/

Approach  : I firstly found the start and the end index of the mismatched or unsorted array (ignoring duplicates). Then I simply traversed through this start to end range and stored the maximum and the minimum element in that range. Then by two more traversal I found the positions of this elements in the actual array. For the minimum element I traversed from left to right and stored it in a variable named 'start' and similarly found the position of the maximum element by traversing from right to left. Then finally returned the result as 'end-start+1'.